修复vip等级问题

+++ app/src/main/java/com/ycf/qianzhihe/section/chat/fragment/BaseChatFragment.java
@@ -1403,8 +1403,9 @@ public class BaseChatFragment extends BaseInitFragment implements EMMessageListe
                 UserComm.getUserInfo().getUserHead());
         message.setAttribute(Constant.NICKNAME,
                 UserComm.getUserInfo().getNickName());
+
         message.setAttribute(Constant.USERVIPLEVEL,
-                NumberUtils.parseInt(UserComm.getUserInfo().getVipLevel()));
+                (int) NumberUtils.parseDouble(UserComm.getUserInfo().getVipLevel()));
 
         if (groupDetailInfo != null && !TextUtils.isEmpty(groupDetailInfo.getGroupUserNickName())) {
             message.setAttribute(Constant.NICKNAME, groupDetailInfo.getGroupUserNickName());
